Add files via upload
I cleaned up the code from file hawaii_test_CF_RR, by removing the redundancies in each time section (adding dates, cumulative summation, etc. to the beginning of the code). 

Also, I simplified the graph compilation code to a single chunk at the end of the .R file. Anyone can stick this at the end of their code and it will work as long as your simulations have the same names as this code. If not, it is a quick fix.